Add basic CorporateActionsService implementation
